1.0 remove sessionstorage error

+++ src/app/components/friend/friend.component.ts
@@ -1,9 +1,10 @@
-import { Component , OnInit} from '@angular/core';
+import {Component, inject, OnInit, PLATFORM_ID} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {NgForOf} from "@angular/common";
+import {isPlatformBrowser, NgForOf} from "@angular/common";
 import {ModalDismissReasons, NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
 export class Friend {
+
   constructor(
     public id: number,
     public username: string,
@@ -26,6 +27,7 @@ export class Friend {
 export class FriendComponent implements OnInit{
   friends: Friend[] | undefined;
   private deleteId: number | undefined;
+  private readonly platformId = inject(PLATFORM_ID);
 
   constructor(
     private httpClient: HttpClient,
@@ -33,14 +35,16 @@ export class FriendComponent implements OnInit{
   ) {
   }
 
-  getUsers(){
-    // @ts-ignore
-    this.httpClient.get<any>('http://localhost:8080/users/'+ sessionStorage.getItem('id')+'/friends').subscribe(
-      response => {
-        console.log(response);
-        this.friends = response;
-      }
-    );
+  getUsers() {
+    if (isPlatformBrowser(this.platformId)) {
+      // @ts-ignore
+      this.httpClient.get<any>('http://localhost:8080/users/' + sessionStorage.getItem('id') + '/friends').subscribe(
+        response => {
+          console.log(response);
+          this.friends = response;
+        }
+      );
+    }
   }
 
   ngOnInit(): void {
@@ -56,11 +60,13 @@ export class FriendComponent implements OnInit{
   }
 
   onDelete() {
-    const deleteURL = 'http://localhost:8080/users/'+ sessionStorage.getItem('id')+'/' + this.deleteId + '/delfriend';
-    this.httpClient.delete(deleteURL)
-      .subscribe((results) => {
-        this.ngOnInit();
-        this.modalService.dismissAll();
-      });
+    if (isPlatformBrowser(this.platformId)) {
+      const deleteURL = 'http://localhost:8080/users/' + sessionStorage.getItem('id') + '/' + this.deleteId + '/delfriend';
+      this.httpClient.delete(deleteURL)
+        .subscribe((results) => {
+          this.ngOnInit();
+          this.modalService.dismissAll();
+        });
+    }
   }
 }
